:recycle: :art: redesign auth package
