+++ api/delete_pipeline.go
@@ -24,7 +24,7 @@ func DeletePipelineUsecase(
 	pipelinesManager *pipelines.Manager,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_WRITE_PIPELINES,
 			func(
