+++ api/delete_transformer.go
@@ -24,7 +24,7 @@ func DeleteTransformerUsecase(
 	pipelinesManager *pipelines.Manager,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_WRITE_TRANSFORMERS,
 			func(
