+++ api/get_pipeline.go
@@ -27,7 +27,7 @@ func GetPipelineUsecase(
 	pipelinesManager *pipelines.Manager,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_READ_PIPELINES,
 			func(
