+++ api/get_transformer.go
@@ -25,7 +25,7 @@ func GetTransformerUsecase(
 	pipelinesManager *pipelines.Manager,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_READ_TRANSFORMERS,
 			func(
