+++ api/list_stream_fields.go
@@ -23,7 +23,7 @@ func ListStreamFieldsUsecase(
 	logDb *logstorage.Storage,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_READ_STREAMS,
 			func(
