+++ api/list_streams.go
@@ -21,7 +21,7 @@ func ListStreamsUsecase(
 	logDb *logstorage.Storage,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_READ_STREAMS,
 			func(
