+++ api/list_transformers.go
@@ -22,7 +22,7 @@ func ListTransformersUsecase(
 	pipelinesManager *pipelines.Manager,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_READ_TRANSFORMERS,
 			func(
