+++ api/save_pipeline.go
@@ -27,7 +27,7 @@ func SavePipelineUsecase(
 	pipelinesManager *pipelines.Manager,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_WRITE_PIPELINES,
 			func(
