+++ api/save_transformer.go
@@ -25,7 +25,7 @@ func SaveTransformerUsecase(
 	pipelinesManager *pipelines.Manager,
 ) usecase.Interactor {
 	u := usecase.NewInteractor(
-		auth.RequireScopeApiMiddleware(
+		auth.RequireScopeApiDecorator(
 			authDb,
 			auth.SCOPE_WRITE_TRANSFORMERS,
 			func(
