+++ web/controllers/main.go
@@ -21,6 +21,8 @@ func MainController(
 ) http.Handler {
 	mux := http.NewServeMux()
 
+	userSys := auth.NewUserSystem(authDb)
+
 	mux.HandleFunc("GET /web/{$}", func(w http.ResponseWriter, r *http.Request) {
 		streamCount := 0
 		transformerCount := 0
@@ -30,7 +32,7 @@ func MainController(
 		notifications := []string{}
 
 		user := auth.GetContextUser(r.Context())
-		scopes, err := authDb.ListUserScopes(user)
+		scopes, err := userSys.ListUserScopes(user.Name)
 		if err != nil {
 			slog.ErrorContext(
 				r.Context(),
